pyproject.toml
[project]
name = "github-actions-integration"
version = "0.1.0"
dependencies = [
    "fastmcp",
]
